Implemented Job Worker for ThirdParty Job API
This job worker can be used by AWS Partners to integrate their products into AWS
CodePipeline and make them available to all AWS customers. The customers can use
the partner tools and products in their software release process.

JobWorkerDaemon implements Apache Common Daemon interface so that it can be run
as a Linux daemon (jsvc) or Windows service (procrun). It schedules the job
poller at a fixed interval to check if there are pending jobs in AWS
CodePipeline. Configure the ActionType to poll for in the JobWorkerConfiguration
class.

The JobPoller calls the CodePipeline API to get the available jobs. It
acknowledges each job and schedules a new thread to start working on the task.
It waits for the result from the JobProcessor and reports back success or
failure to the JobService.

JobService is simply a wrapper interface around the CodePipeline Job API.

Implement your custom logic to integrate with your products and tools in the
JobProcessor. Also override the hard-coded client token in the
DefaultClientTokenProvider and lookup the correct value passed in the connect UI
flow.

The job worker uses the AWS SDK for Java to make calls to the CodePipeline API.
Use EC2 instance roles or provide IAM credentials to the AWS SDK.

Created deployment scripts so that the job worker can be easily installed on EC2
or on-premise hosts using AWS CodeDeploy.