Added sample implementation for job processor
The job processor makes a job fail if the action configuration contains a
property 'JobStatus' with the value 'JobFailed'.

This implementation can be used to test the success or failure of the jobs and
shows how the job result can be reported back to AWS CodePipeline.

If a job succeeds, AWS CodePipeline rolls out this change to the next action or
stage. If a job fails, the change is not rolled out any further and the pipeline
execution stops.