Added Custom Action Support
Extended the job worker to support the custom action AWS CodePipeline Job API.
Custom Actions allow AWS Customers to integrate their on-premise systems into
AWS CodePipeline and make them part of their release pipelines.

Created a second implementation of the JobService interface. The
CustomActionJobService uses the Custom Action Job APIs to poll for jobs,
acknowledge them and report back status.

The configuration class can be passed as an argument to the daemon so that the
job worker can be started with different configurations
(ThirdPartyJobWorkerConfiguration and CustomActionJobWorkerConfiguration).