Updated to AWS SDK for Java 1.10.5.1
Updated the Java SDK to version 1.10.5.1.

Removed constructor overload of the FailureDetails class. The failure message
is now required and has to be always provided.

Using AWSCodePipeline interface instead of depending on the actual client
implementation.

Added JobDataGenerator for unit tests and validating that the artifacts,
session credentials and pipeline context are properly mapped.